fixed konsave call (during the install)
while installing the profile, sourcing the 'aliasrc' file does not enable the newly added alias. therefore, that line has been removed, and the 'konsave' call uses the absolute path.

in addition, the 'do-thru-ui.md' file was 'cat' twice because of a duplicate line in the 'konsave' module. it has also been removed.